pyproject.toml
[project]
name = "deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.6.0",
    "codetiming>=1.4.0",
    "datasets>=3.6.0",
    "dill>=0.3.8",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.21",
    "loguru>=0.7.3",
    "numpy>=1.26.4",
    "openai>=1.78.1",
    "pandas>=2.2.3",
    "polars>=1.29.0",
    "pyarrow>=15.0.0",
    "pydantic>=2.11.4",
    "pylatexenc>=2.10",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "ruff>=0.11.8",
    "tensordict<=0.6.2",
    "transformers>=4.40.0",
    "trl>=0.17.0",
    "vllm==0.8.4",
    "wandb>=0.19.11",
    "peft",
    "pybind11",
    "ray[default]>=2.10",
    "torchdata",
    "base58",
    "nltk",
    "PyYAML",
    "rank_bm25",
    "rouge",
    "spacy",
    "tiktoken",
    "torch",
    "tqdm",
    "bm25s[core]==0.2.0",
    "fschat",
    "streamlit",
    "chonkie>=0.4.0",
    "gradio>=5.0.0",
    "rouge-chinese",
    "jieba",
    "flash-attn>=2.7.4.post1",
    "tensorboard>=2.19.0",
]

[tool.ruff]
exclude = ["third_party"]
